mvp/backend/pyproject.toml
[tool.poetry]
name = "saas-ia-mvp"
version = "2.6.0"
description = "SaaS IA MVP - Plateforme modulaire d'intelligence artificielle"
authors = ["SaaS-IA Team"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
# FastAPI & Server
fastapi = "^0.135.1"
uvicorn = {extras = ["standard"], version = "^0.42.0"}

# Database
sqlmodel = "^0.0.37"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.31.0"
aiosqlite = "^0.22.1"
alembic = "^1.18.4"

# Cache & Queue
redis = "^5.2.1"
celery = {extras = ["redis"], version = "^5.6.2"}
flower = "^2.0.1"

# Payments
stripe = "^8.0.0"

# Validation & Settings
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.13.1"
email-validator = "^2.1.0"

# Auth
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.2.1"  # Pin bcrypt to compatible version with passlib
python-multipart = "^0.0.22"

# AI Services
assemblyai = "^0.58.0"
yt-dlp = "^2026.3.17"
crawl4ai = ">=0.4.0"

# AI Providers
google-generativeai = "^0.8.6"
anthropic = "^0.86.0"
openai = "^1.54.5"  # Pinned v1.x for Groq compatibility

# Utils
python-dotenv = "^1.2.2"
structlog = "^25.5.0"
python-dateutil = "^2.9.0"

# Monitoring & Observability
prometheus-client = "^0.24.1"
pyyaml = "^6.0.3"

# Rate Limiting
slowapi = "^0.1.9"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-cov = "^7.1.0"
httpx = "^0.28.1"

# Linting & Formatting
ruff = "^0.15.7"
mypy = "^1.19.1"
black = "^26.3.1"

# Security
bandit = "^1.9.4"

[tool.ruff]
line-length = 120
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term --cov-report=html"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
